fix(stella-pipeline): end a verdict negator's reach at its clause
#1808 narrowed the negation window from three tokens to two, which fixed
the reported "Not a problem. PASS". But the window measures the wrong
thing, and one shape survives it:

    "No issues. Not blocking. PASS"  ->  FAIL

The negator there sits ONE token from the verdict — nearer than "cannot
currently pass", which must keep negating — so no window wide enough for
the second can exclude the first. What separates them is not distance but
the sentence between them.

So the negator's scope now ends at terminal punctuation (`. ! ? ; :`) and
the two-token window still applies *within* a clause, where unpunctuated
prose has nothing else to stop it. Both bounds are load-bearing: drop the
clause break and "No issues. Not blocking. PASS" inverts; drop the window
and "Not a problem PASS" inverts.

A comma and an em-dash are deliberately not boundaries — they join clauses
at least as often as they separate them ("it does not, in this case,
pass"), so counting them would drop real negations to buy back cases the
window already handles.

The direction of this bug is the expensive one: an approval read as a
refusal sends a finished, correct candidate back for revision and spends a
whole turn's budget denying its own result.

`verify/tests.rs` had reached the 1500-line ratchet exactly, so the
verdict-parsing tests move to a `parse` submodule beside `uncollected`
rather than growing the parent over the limit.

Refs #1798
